requirements.txt
duckdb==1.2.1
duckdb_engine==0.17.0
firecrawl-py==1.15.0
langchain-openai==0.3.14
pydantic-settings==2.9.1
beautifulsoup4==4.13.3
lxml==5.3.2
langchain-community==0.3.21
streamlit==1.44.1